+++ CHANGELOG.md
@@ -53,6 +53,8 @@
 * Implement easy way to copy Barcode in the Item record. Refs UIIN-1250.
 * Implement easy way to copy HRID in the Holdings record. Refs UIIN-1001.
 * Implement easy way to copy HRID in the Instance record. Refs UIIN-1000.
+* Add Holdings sources to settings. Refs UIIN-1229.
+
 
 ## [4.0.1](https://github.com/folio-org/ui-inventory/tree/v4.0.1) (2020-07-01)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v4.0.0...v4.0.1)